pyproject.toml
[project]
name = "bidauto-post-generator-service"
version = "0.1.0"
description = ""
authors = [
    {name = "peyro",email = "peyrovskaaa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "aio-pika (>=9.5.7,<10.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "grpcio (>=1.74.0,<2.0.0)",
    "protobuf (>=6.32.1,<7.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.21.0"

